refactor repository for dto
