+++ src/main/java/co/jht/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package co.jht.repository;
 
-import co.jht.entity.AppUser;
+import co.jht.model.domain.persist.entity.appuser.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<AppUser, Long> {
